Добавлено получение информации о полётах через api и вывод рейсов клиенту

+++ src/main/java/com/aerosales/aviaticketsspring/repos/UsersRepository.java
@@ -3,6 +3,6 @@ package com.aerosales.aviaticketsspring.repos;
 import com.aerosales.aviaticketsspring.pojo.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRepository extends JpaRepository<Users, Long> {
+public interface UsersRepository extends JpaRepository<Users, Integer> {
     Users findByUsername(String username);
 }
